+++ lib/questionnaires/model/expression/src/resource_expression_evaluator.dart
@@ -15,6 +15,5 @@ class ResourceExpressionEvaluator extends ExpressionEvaluator {
   ResourceExpressionEvaluator(
     String name,
     this.resourceBuilder,
-    Iterable<ExpressionEvaluator> upstreamExpressions,
-  ) : super(name, upstreamExpressions);
+  ) : super(name, []);
 }